HOL-Light: Add support for MacOS
The HOL-Light/s2n-bignum proofs expect ELF formatted object files,
but compilation on MacOS produces Mach-O object files.

This commit modifies the makefile for the HOL-Light proofs in mlkem-native
to explicitly convert from Mach-O to ELF using gobjcopy.

Unfortunately, there does not seem to be a way to get a version of
`objcopy` through `nix` that supports the required conversion, hence
we need to fall back to documentation and expressive error messages
to nudge the user to manually install `gobjcopy` via `brew install binutils`
when it cannot be found.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>